+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fixes to asusctl CLI tool to show fan curves
 - Fixes to asusd to ensure fan curve defaults are loaded if the config file fails
+- Further refine the asusctl CLI for fan-curve control
 
 ### Added
 - Support for GV601V LED modes